Add timeout for fade in effect
When a surface is created, a fade in effect occurs (if specified in the
arguments). This animation is desirable when swaylock is initially run,
however, it is undesirable after it is run.

For example, on a running swaylock process, if the surface is destroyed
when a laptop is closed, the surface must be recreated when the lid is
reopened and the output turns on again. During this recreation, the fade
in animation replays, temporarily showing the working state of the
machine long after swaylock was initially run.

This has obvious security and privacy concerns - anyone can see the
working state of a machine by closing and reopening a laptop lid, or by
hooking up another monitor and forcing another surface to be created.
This PR is a simple attempt to fix that by only allowing the fade
animation to occur in a small time window immediately after swaylock
starts.

Resolves #12