refactor: Apply the fade-in effect using screenshots
With ext_session_lock_v1, it is no longer possible to have an lockscreen
surface with alpha channel. Instead, we have to implement the alpha
blending ourselves.

Instead of reinventing the wheel yet again, we simply make use of the
screenshot we have taken beforehand (before the lock was even
initiated). We draw our intended background image on top of an
unmodified version of the screenshots directly, by using
cairo_paint_with_alpha().