Apply effects before requesting lock
This reduces flickering while the effects are being applied. The
downside is that now there might be a small delay between running
swaylock and when the lock is actually applied, but that should be fine
given there are visual confirmations for the lock (as in, whether
swaylock showed up or not).