Render before sending ACK for ext_session_lock_v1
This further reduces the blank time between activating the lock and we
rendering the first frame. Note that this means we must do the commit
outside of render_frame_background().

render_frame_background() itself was modified with a new parameter.
Compared to just removing wl_surface_commit() from it, this approach can
reduce the chance of future mismerges by forgetting wl_surface_commit().